freecodecamp/relational database/pyproject.toml
[tool.poetry]
name = "relational database"
version = "0.1.0"
description = ""
authors = ["MorphZG <zoran1607@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2 = "^2.9.5"
python-dotenv = "^0.21.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
